+++ R/itemfit.R
@@ -88,10 +88,10 @@
 #'   to approximate the scaling factor for X2* as well as the scaled degrees of freedom estimates
 #' @param ETrange rangone of integration nodes for Stone's X2* statistic
 #' @param ETpoints number of integration nodes to use for Stone's X2* statistic
-#' @param impute a number indicating how many imputations to perform (passed to
-#'   \code{\link{imputeMissing}}) when there are missing data present.
-#'   Will return a data.frame object with the mean estimates
-#'   of the stats and their imputed standard deviations
+# @param impute a number indicating how many imputations to perform (passed to
+#   \code{\link{imputeMissing}}) when there are missing data present.
+#   Will return a data.frame object with the mean estimates
+#   of the stats and their imputed standard deviations
 #' @param par.strip.text plotting argument passed to \code{\link{lattice}}
 #' @param par.settings plotting argument passed to \code{\link{lattice}}
 #' @param ... additional arguments to be passed to \code{fscores()} and \code{\link{lattice}}
@@ -177,7 +177,7 @@
 #' # itemfit(x, 'X2*') # Stone's 1993 statistic
 #' # itemfit(x, 'X2*_df') # Stone's 2000 scaled statistic with df estimate
 #'
-#' #empirical tables
+#' #empirical tables for X2 statistic
 #' itemfit(x, empirical.table=1)
 #' itemfit(x, empirical.table=21)
 #'
@@ -217,7 +217,7 @@
 #' mod2 <- mirt(dat, 1, 'Rasch')
 #' itemfit(mod2, 'infit', method = 'ML')
 #'
-#' #massive list of tables
+#' #massive list of tables for S-X2
 #' tables <- itemfit(mod, S_X2.tables = TRUE)
 #'
 #' #observed and expected total score patterns for item 1 (post collapsing)
@@ -228,11 +228,7 @@
 #' dat[sample(1:prod(dim(dat)), 100)] <- NA
 #' raschfit <- mirt(dat, 1, itemtype='Rasch')
 #'
-#' # use imputation if the proportion of missing data is relatively small
-#' mirtCluster() # run in parallel
-#' itemfit(raschfit, c('S_X2', 'infit'), impute = 10)
-#'
-#' #alternative route: use only valid data by removing rows with missing terms
+#' #use only valid data by removing rows with missing terms
 #' itemfit(raschfit, c('S_X2', 'infit'), na.rm = TRUE)
 #'
 #' # note that X2, G2, PV-Q1, and X2* do not require complete datasets
@@ -249,20 +245,21 @@ itemfit <- function(x, fit_stats = 'S_X2', which.items = 1:extract.mirt(x, 'nite
                     pv_draws = 30, boot = 1000, boot_dfapprox = 200,
                     ETrange = c(-2,2), ETpoints = 11,
                     empirical.plot = NULL, empirical.CI = .95, empirical.table = NULL,
-                    empirical.poly.collapse = FALSE, method = 'EAP', Theta = NULL, impute = 0,
+                    empirical.poly.collapse = FALSE, method = 'EAP', Theta = NULL, #impute = 0,
                     par.strip.text = list(cex = 0.7),
                     par.settings = list(strip.background = list(col = '#9ECAE1'),
                                         strip.border = list(col = "black")), ...){
 
-    fn <- function(ind, Theta, obj, vals, ...){
-        tmpobj <- obj
-        tmpdat <- imputeMissing(obj, Theta[[ind]], warn=FALSE)
-        tmpmod <- mirt(tmpdat, model=1, TOL=NA,
-                       technical=list(customK=obj@Data$K, message=FALSE, warn=FALSE))
-        tmpobj@Data <- tmpmod@Data
-        whc <- 1L:length(Theta)
-        return(itemfit(tmpobj, Theta=Theta[[sample(whc[-ind], 1L)]], ...))
-    }
+    impute <- 0
+    # fn <- function(ind, Theta, obj, vals, ...){
+    #     tmpobj <- obj
+    #     tmpdat <- imputeMissing(obj, Theta[[ind]], warn=FALSE)
+    #     tmpmod <- mirt(tmpdat, model=1, TOL=NA,
+    #                    technical=list(customK=obj@Data$K, message=FALSE, warn=FALSE))
+    #     tmpobj@Data <- tmpmod@Data
+    #     whc <- 1L:length(Theta)
+    #     return(itemfit(tmpobj, Theta=Theta[[sample(whc[-ind], 1L)]], ...))
+    # }
     PV_itemfit <- function(mod, which.items = 1:extract.mirt(mod, 'nitems'),
                            draws = 100, ...){
         pv <- fscores(mod, plausible.draws = draws, ...)
@@ -421,7 +418,7 @@ itemfit <- function(x, fit_stats = 'S_X2', which.items = 1:extract.mirt(x, 'nite
     if(na.rm) x <- removeMissing(x)
     if(any(is.na(x@Data$data)) && (Zh || S_X2 || infit) && impute == 0)
         stop('Only X2, G2, PV_Q1, PV_Q1*, X2*, and X2*_df can be computed with missing data.
-             Consider using imputed datasets or passing na.rm=TRUE', call.=FALSE)
+             Pass na.rm=TRUE to remove missing data row-wise', call.=FALSE)
 
     if(is(x, 'MultipleGroupClass') || is(x, 'DiscreteClass')){
         discrete <- is(x, 'DiscreteClass')
@@ -460,41 +457,39 @@ itemfit <- function(x, fit_stats = 'S_X2', which.items = 1:extract.mirt(x, 'nite
         discrete <- TRUE
         pis <- extract.mirt(x, 'pis')
     }
-    if(impute != 0 && !is(x, 'MultipleGroupClass')){
-        if(impute == 0)
-            stop('Fit statistics cannot be computed when there are missing data. Pass a suitable
-                 impute argument to compute statistics following multiple data
-                 imputations', call.=FALSE)
-        if(sum(is.na(x@Data$data)) / length(x@Data$data) > .10)
-            warning('Imputations for large amounts of missing data may be overly conservative. Use with caution', call.=FALSE)
-        stopifnot(impute > 1L)
-        if(is.null(Theta))
-            Theta <- fscores(x, plausible.draws = impute, method = ifelse(method == 'MAP', 'MAP', 'EAP'), ...)
-        collect <- vector('list', impute)
-        vals <- mod2values(x)
-        vals$est <- FALSE
-        collect <- myLapply(1L:impute, fn, Theta=Theta, obj=x, vals=vals, fit_stats=fit_stats,
-                            group.size=group.size, group.bins=group.bins,
-                            mincell=mincell, mincell.X2=mincell.X2,
-                            S_X2.tables=S_X2.tables, empirical.plot=empirical.plot,
-                            empirical.CI=empirical.CI, empirical.table=empirical.table,
-                            method=method, impute=0, discrete=discrete, ...)
-        ave <- SD <- collect[[1L]]
-        pick1 <- 1:nrow(ave)
-        pick2 <- sapply(ave, is.numeric)
-        ave[pick1, pick2] <- SD[pick1, pick2] <- 0
-        for(i in seq_len(impute))
-            ave[pick1, pick2] <- ave[pick1, pick2] + collect[[i]][pick1, pick2]
-        ave[pick1, pick2] <- ave[pick1, pick2]/impute
-        for(i in seq_len(impute))
-            SD[pick1, pick2] <- SD[pick1, pick2] + (ave[pick1, pick2] - collect[[i]][pick1, pick2])^2
-        SD[pick1, pick2] <- sqrt(SD[pick1, pick2]/impute)
-        SD$item <- paste0('SD_', SD$item)
-        SD <- rbind(NA, SD)
-        ret <- rbind(ave, SD)
-        class(ret) <- c('mirt_df', 'data.frame')
-        return(ret)
-    }
+    # if(impute != 0 && !is(x, 'MultipleGroupClass')){
+    #     if(impute == 0)
+    #         stop('Fit statistics cannot be computed when there are missing data. Pass na.rm=TRUE to remove cases row-wise', call.=FALSE)
+    #     if(sum(is.na(x@Data$data)) / length(x@Data$data) > .10)
+    #         warning('Imputations for large amounts of missing data may be overly conservative. Use with caution', call.=FALSE)
+    #     stopifnot(impute > 1L)
+    #     if(is.null(Theta))
+    #         Theta <- fscores(x, plausible.draws = impute, method = ifelse(method == 'MAP', 'MAP', 'EAP'), ...)
+    #     collect <- vector('list', impute)
+    #     vals <- mod2values(x)
+    #     vals$est <- FALSE
+    #     collect <- myLapply(1L:impute, fn, Theta=Theta, obj=x, vals=vals, fit_stats=fit_stats,
+    #                         group.size=group.size, group.bins=group.bins,
+    #                         mincell=mincell, mincell.X2=mincell.X2,
+    #                         S_X2.tables=S_X2.tables, empirical.plot=empirical.plot,
+    #                         empirical.CI=empirical.CI, empirical.table=empirical.table,
+    #                         method=method, impute=0, discrete=discrete, ...)
+    #     ave <- SD <- collect[[1L]]
+    #     pick1 <- 1:nrow(ave)
+    #     pick2 <- sapply(ave, is.numeric)
+    #     ave[pick1, pick2] <- SD[pick1, pick2] <- 0
+    #     for(i in seq_len(impute))
+    #         ave[pick1, pick2] <- ave[pick1, pick2] + collect[[i]][pick1, pick2]
+    #     ave[pick1, pick2] <- ave[pick1, pick2]/impute
+    #     for(i in seq_len(impute))
+    #         SD[pick1, pick2] <- SD[pick1, pick2] + (ave[pick1, pick2] - collect[[i]][pick1, pick2])^2
+    #     SD[pick1, pick2] <- sqrt(SD[pick1, pick2]/impute)
+    #     SD$item <- paste0('SD_', SD$item)
+    #     SD <- rbind(NA, SD)
+    #     ret <- rbind(ave, SD)
+    #     class(ret) <- c('mirt_df', 'data.frame')
+    #     return(ret)
+    # }
     if(S_X2.tables || discrete) Zh <- X2 <- FALSE
     ret <- data.frame(item=colnames(x@Data$data)[which.items])
     itemloc <- x@Model$itemloc