+++ NEWS.md
@@ -1,5 +1,8 @@
 # Changes in mirt 1.31
 
+- `impute` argument in `itemfit()` and `M2()` have been deprecated in favour of removing data
+  row-wise via `na.rm=TRUE`
+
 - Acceptance ratio when using MH samplers now retuned prior to 'Stage 2' during estimation so that 
   these ratios are better behaved. As well, an heuristic improved method for increasing/decreasing
   the acceptance ratios is now implemented